pdf2epub/requirements.txt
pymupdf
ebooklib
arabic-reshaper
python-bidi

# Optional OCR dependencies
# pytesseract
# pillow
